Add option to configure and build without graphics
Make the configuration and build of graphical demos and
their dependencies (Magnum, Corrade, SDL2) depend on a
new CMake cache variable ENABLE_GRAPHICS. Auto-configure
this new variable based on the result of find_package(SDL2).

A warning is emitted if ENABLE_GRAPHICS is unset and SDL2 is
not found. The warning can be silenced by explicitly setting
the option to OFF. In case the option is explicitly set to ON
and SDL2 is not found, a fatal error is emitted.

Add an explanation of the new option in the README.

Add a new CI step building the no-graphics version. Set
ENABLE_GRAPHICS=ON explicitly in the baseline CI build.