Update clang-format include order rules (#38)
We should be including the sycl.hpp header from the sycl/ folder.
Changing from CL/sycl to sycl/sycl started forcing some header file
reordering, which caused pre-commit failures in the CI. So I have
changed the clang-format rules to better express what we want, which
has then caused other formatting changes for files that were not already
compliant to my dread will.

* all "headers.hpp"
* <third_party/includes.h>
* <sycl/sycl.hpp>
* <cppstdlibheaders>

Other than that, changing the matmul header caused the `cl::` namespace
to disappear, so was removed. Other formatting in the tuple_utils header
started becoming a bit nasty so knowing that SYCL requires C++ 17, I
changed to `decltype(auto)` for some of the function return types.
This is a bit more terse and functions exactly the same as the previous
version.